# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Risk Radar",
    "description": "Real-time transaction risk intelligence. **Advisory**: this service produces a decision (ALLOW / MONITOR / REVIEW / HOLD) with its reasons. It does not block, hold, reverse or freeze anything \u2014 enforcement is the caller's responsibility and is documented as such in the contract.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/cases/{case_id}/submissions": {
      "post": {
        "tags": [
          "fraud approval"
        ],
        "summary": "Submit",
        "operationId": "submit_v1_cases__case_id__submissions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FraudSubmissionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Submit V1 Cases  Case Id  Submissions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "fraud approval"
        ],
        "summary": "History",
        "operationId": "history_v1_cases__case_id__submissions_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response History V1 Cases  Case Id  Submissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/approvals": {
      "get": {
        "tags": [
          "fraud approval"
        ],
        "summary": "Queue",
        "description": "Waiting proposals, never the reader's own: a lead's queue is other people's work.",
        "operationId": "queue_v1_approvals_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Queue V1 Approvals Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/submissions/{submission_id}/decision": {
      "post": {
        "tags": [
          "fraud approval"
        ],
        "summary": "Decide",
        "operationId": "decide_v1_submissions__submission_id__decision_post",
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FraudDecisionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Decide V1 Submissions  Submission Id  Decision Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FraudDecisionIn": {
        "properties": {
          "decision": {
            "type": "string",
            "enum": [
              "APPROVE",
              "REJECT",
              "RETURN"
            ],
            "title": "Decision"
          },
          "reason": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 10,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "decision",
          "reason"
        ],
        "title": "FraudDecisionIn",
        "description": "D93: a lead's decision on somebody else's proposal."
      },
      "FraudSubmissionIn": {
        "properties": {
          "proposed_outcome": {
            "type": "string",
            "enum": [
              "CONFIRMED_FRAUD",
              "FALSE_POSITIVE",
              "INCONCLUSIVE"
            ],
            "title": "Proposed Outcome"
          },
          "rationale": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 20,
            "title": "Rationale"
          },
          "restrictions": {
            "items": {
              "$ref": "#/components/schemas/RestrictionIn"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Restrictions",
            "default": []
          },
          "expected_case_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Case Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "proposed_outcome",
          "rationale"
        ],
        "title": "FraudSubmissionIn",
        "description": "D93: the analyst's proposal. It decides nothing until a lead approves it."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RestrictionIn": {
        "properties": {
          "action": {
            "type": "string",
            "enum": [
              "DEBIT_RESTRICTION",
              "CHANNEL_RESTRICTION",
              "CARD_FREEZE",
              "BENEFICIARY_RESTRICTION"
            ],
            "title": "Action"
          },
          "account_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Token"
          },
          "beneficiary_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Beneficiary Token"
          },
          "channel": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "MOBILE_APP",
                  "WEB",
                  "USSD",
                  "POS",
                  "ATM",
                  "AGENT",
                  "BRANCH",
                  "API"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Channel"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action"
        ],
        "title": "RestrictionIn",
        "description": "D93: one action a submission asks the bank to take, from a fixed list."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}